Fix typo
In the ```Devless is a tool for ready-made back-end for development of web or mobile applications. It is fully open source under the permissive Apache v2 license. This means that you can develop your front end without worrying about neither back-end code or the business risk of a *propitiatory* backend-as-a-service.``` . Changed `propitiatory` to `proprietary`.